refactor: change not-found folder
